requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
